Compatibility for redmine 5.x
Add Spanish translation
